Move test-only methods from LocalDateTimeUtils to TestPreparedStatement
